requirements.txt
requests
python-dotenv
playwright==1.44.0